# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdverTest API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/experiments/{experiment_id}/runs": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "List Experiment Runs",
        "operationId": "list_experiment_runs_experiments__experiment_id__runs_get",
        "security": [
          {
            "userSession": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunView"
                  },
                  "title": "Response List Experiment Runs Experiments  Experiment Id  Runs Get"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu phiên hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Thiếu permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-permission": "experiment.read"
      }
    },
...
    "/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_runs__run_id__get",
        "security": [
          {
            "userSession": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu phiên hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Thiếu permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-permission": "experiment.read"
      }
    },
...
    "/internal/worker/lease": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Lease",
        "description": "Nhận experiment `queued` cũ nhất của target (hoặc experiment có lease đã hết hạn).",
        "operationId": "lease_internal_worker_lease_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerLease"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Không có experiment nào chờ"
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ]
      }
    },
    "/internal/worker/experiments/{experiment_id}/bundle": {
      "get": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Get Bundle",
        "description": "Mọi thứ cần để chạy experiment, kèm presigned URL tải (hết hạn sau 15 phút).",
        "operationId": "get_bundle_internal_worker_experiments__experiment_id__bundle_get",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerJobBundle"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/worker/heartbeat": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Heartbeat",
        "description": "Gia hạn lease 60 giây.",
        "operationId": "heartbeat_internal_worker_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerDirective"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ]
      }
    },
    "/internal/worker/runs/{run_id}/start": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Start Run",
        "description": "Chạy run, hoặc bỏ qua khi đã có run `completed` cùng fingerprint.",
        "operationId": "start_run_internal_worker_runs__run_id__start_post",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStartResponse"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/worker/runs/{run_id}/progress": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Report Progress",
        "description": "Báo tiến độ sau mỗi batch; cộng dồn thời gian xử lý.",
        "operationId": "report_progress_internal_worker_runs__run_id__progress_post",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressReport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerDirective"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/worker/runs/{run_id}/artifact-url": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Artifact Url",
        "description": "Presigned URL (PUT, GET, DELETE) cho một khóa nằm trong `runs/<run_id>/`.",
        "operationId": "artifact_url_internal_worker_runs__run_id__artifact_url_post",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactUrlResponse"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/worker/runs/{run_id}/complete": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Complete Run",
        "description": "Kết quả cuối của run kèm failure case.",
        "operationId": "complete_run_internal_worker_runs__run_id__complete_post",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCompletion"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/worker/runs/{run_id}/skip": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Skip Run",
        "description": "Bỏ run `queued` do dừng sớm (Phase 6); `trigger_run_id` là run cùng attack đã làm model\nsụp.",
        "operationId": "skip_run_internal_worker_runs__run_id__skip_post",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSkipRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/worker/runs/{run_id}/patch": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Register Patch",
        "description": "Đăng ký patch vừa train xong cho run patch (Phase 6). Khóa đã có thì giữ bản cũ và trả bản\nđó.",
        "operationId": "register_patch_internal_worker_runs__run_id__patch_post",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRegistration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchArtifact"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/worker/cost-profiles": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Submit Cost Profile",
        "description": "Cost profile đo bằng calibration cho target của token.",
        "operationId": "submit_cost_profile_internal_worker_cost_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ]
      }
    },
    "/internal/worker/experiments/{experiment_id}/runs": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Create Search Run",
        "description": "Tạo run `queued` cho điểm tìm ngưỡng kế tiếp (Phase 7). Vi phạm (attack không ở chế độ tìm\nngưỡng, level ngoài `[lo, hi]`, vượt `max_points`) trả `422`, không `409`.",
        "operationId": "create_search_run_internal_worker_experiments__experiment_id__runs_post",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRunCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleRun"
                }
              }
            }
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/worker/experiments/{experiment_id}/search-result": {
      "post": {
        "tags": [
          "internal-worker"
        ],
        "summary": "Submit Search Result",
        "description": "SearchResult tạm thời sau mỗi điểm, hoặc kết quả cuối (Phase 7).",
        "operationId": "submit_search_result_internal_worker_experiments__experiment_id__search_result_post",
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchResultReport"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "501": {
            "description": "Chưa cài đặt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error (sai schema) hoặc invalid_request (sai nghiệp vụ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Body lỗi thống nhất của mọi endpoint: {\"error\": {\"code\", \"message\"}}."
      },
...
      "HeartbeatRequest": {
        "properties": {
          "lease_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lease Id"
          },
          "experiment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Experiment Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lease_id",
          "experiment_id"
        ],
        "title": "HeartbeatRequest"
      },
...
      "WorkerLease": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "experiment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Experiment Id"
          },
          "lease_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lease Id",
            "description": "Đổi mỗi lần lease; gửi kèm mọi request sau đó"
          },
          "lease_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Lease Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "experiment_id",
          "lease_id",
          "lease_expires_at"
        ],
        "title": "WorkerLease",
        "description": "Trả về từ `POST /lease` (không có job thì `204`)."
      }
    },
    "securitySchemes": {
      "userSession": {
        "type": "apiKey",
        "description": "Cookie phiên của người dùng (token ngẫu nhiên, httpOnly; phiên lưu phía server).",
        "in": "cookie",
        "name": "advertest_session"
      },
      "workerToken": {
        "type": "http",
        "description": "Token của compute target, chỉ dùng cho /internal/worker.",
        "scheme": "bearer"
      }
    }
  }
}